+++ Source/GPU/Add On - DirectX/CDirectXShader.cpp
@@ -72,8 +72,8 @@ void CDirectXVertexShader::setup(ID3D11Device& d3dDevice, ID3D11DeviceContext3& 
 		// Create Vertex Shader
 		HRESULT	result;	
 		result =
-				d3dDevice.CreateVertexShader(data.getValue().getBytePtr(), data.getValue().getByteCount(), NULL,
-						&mInternals->mShader);
+				d3dDevice.CreateVertexShader(data.getValue().getBytePtr(), (SIZE_T) data.getValue().getByteCount(),
+						NULL, &mInternals->mShader);
 		AssertFailIf(FAILED(result));
 
 		// Create resources
@@ -156,8 +156,8 @@ void CDirectXPixelShader::setup(ID3D11Device& d3dDevice, ID3D11DeviceContext3& d
 
 		// Create Pixel Shader
 		HRESULT	result =
-						d3dDevice.CreatePixelShader(data.getValue().getBytePtr(), data.getValue().getByteCount(), NULL,
-								&mInternals->mShader);
+						d3dDevice.CreatePixelShader(data.getValue().getBytePtr(),
+								(SIZE_T) data.getValue().getByteCount(), NULL, &mInternals->mShader);
 		AssertFailIf(FAILED(result));
 
 		// Create resources
